requirements.txt
mcpi
tqdm
wheel
PyYAML
setuptools
opencv-python